requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic_settings==2.4.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
pandas==2.2.2
fpdf2==2.7.9
python-multipart==0.0.9
pytest==8.3.2
pytest-asyncio==0.23.8
python-jose[cryptography]
email-validator
